% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg.R
\name{cg_coin}
\alias{cg_coin}
\title{Detailed coin information}
\usage{
cg_coin(coin)
}
\arguments{
\item{coin}{CoinGecko coin ID}
}
\value{
tibble: one row with id, symbol, name, description, categories,
genesis_date, market_data (nested)
}
\description{
Detailed coin information
}
