% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg.R
\name{cg_trending}
\alias{cg_trending}
\title{Trending coins (top 15 by search volume)}
\usage{
cg_trending()
}
\value{
tibble: id, name, symbol, market_cap_rank, score
}
\description{
Trending coins (top 15 by search volume)
}
