% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg.R
\name{cg_ohlc}
\alias{cg_ohlc}
\title{Historical OHLC candlestick data}
\usage{
cg_ohlc(coin, vs_currency = "usd", days = 30)
}
\arguments{
\item{coin}{CoinGecko coin ID}

\item{vs_currency}{Quote currency (default "usd")}

\item{days}{Days back: 1, 7, 14, 30, 90, 180, 365, "max"}
}
\value{
tibble: timestamp, open, high, low, close, coin
}
\description{
Historical OHLC candlestick data
}
